+++ src/content/main.js
@@ -5,6 +5,7 @@ import { watchUrlChanges } from "./url-watcher.js";
 import { watchKetcherDialog } from "./overlay-watcher.js";
 import { ensurePanel, renderFromState, initSamplePanelFields } from "./features/sample-panel.js";
 import { ensurePrintButtons } from "./features/print-buttons.js";
+import { initSynonymEnrichment } from "./features/synonym-enrichment.js";
 import {
   ensureDepletedStyle,
   markDepletedSamplesInSelector,
@@ -85,6 +86,7 @@ function init() {
   injectPageScript();
   initDoseResponseOverride();
   initSamplePanelFields();
+  initSynonymEnrichment();
 
   window.addEventListener("message", handleMessage);